+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/ac/AcLoadFlowLogger.java
@@ -6,6 +6,7 @@
  */
 package com.powsybl.loadflow.simple.ac;
 
+import com.powsybl.loadflow.simple.ac.observer.DefaultAcLoadFlowObserver;
 import com.powsybl.loadflow.simple.equations.Equation;
 import com.powsybl.loadflow.simple.equations.EquationSystem;
 import org.slf4j.Logger;
