+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/ac/AcLoadFlowResult.java
@@ -6,6 +6,8 @@
  */
 package com.powsybl.loadflow.simple.ac;
 
+import com.powsybl.loadflow.simple.ac.nr.NewtonRaphsonStatus;
+
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
