+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/ac/AcloadFlowEngine.java
@@ -10,6 +10,10 @@ import com.google.common.base.Stopwatch;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.loadflow.LoadFlowParameters;
 import com.powsybl.loadflow.simple.ac.equations.AcEquationSystem;
+import com.powsybl.loadflow.simple.ac.nr.NewtonRaphson;
+import com.powsybl.loadflow.simple.ac.nr.NewtonRaphsonParameters;
+import com.powsybl.loadflow.simple.ac.nr.NewtonRaphsonResult;
+import com.powsybl.loadflow.simple.ac.observer.AcLoadFlowObserver;
 import com.powsybl.loadflow.simple.equations.EquationContext;
 import com.powsybl.loadflow.simple.equations.EquationSystem;
 import com.powsybl.loadflow.simple.network.NetworkContext;
