+++ simple-loadflow/src/main/java/com/powsybl/loadflow/simple/ac/nr/NewtonRaphsonParameters.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.loadflow.simple.ac;
+package com.powsybl.loadflow.simple.ac.nr;
 
 import com.powsybl.loadflow.LoadFlowParameters;
